pyproject.toml
[project]
name = "sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "chroma>=0.2.0",
    "faiss-cpu>=1.11.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-core>=0.3.58",
    "langchain-groq>=0.3.2",
    "langchain-openai>=0.3.16",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "streamlit>=1.45.0",
    "tiktoken>=0.9.0",
]
